+++ python/src/scipp/_variable.py
@@ -163,6 +163,76 @@ def empty(*,
     return _cpp.empty(dims, shape, unit, dtype, variances)
 
 
+def _to_eigen_layout(a):
+    # Numpy and scipp use row-major, but Eigen matrices use column-major,
+    # transpose matrix axes for copying values.
+    return _np.moveaxis(a, -1, -2)
+
+
+def matrix(*,
+           unit: _Union[_cpp.Unit, str] = _cpp.units.dimensionless,
+           value: _Union[_np.ndarray, list]):
+    """Constructs a zero dimensional :class:`Variable` holding a single 3x3
+    matrix.
+
+    :seealso: :py:func:`scipp.matrices`
+
+    :param value: Initial value, a list or 1-D numpy array.
+    :param unit: Optional, unit. Default=dimensionless
+    :returns: A scalar (zero-dimensional) Variable.
+    :rtype: Variable
+    """
+    return _cpp.matrices(dims=[], unit=unit, values=_to_eigen_layout(value))
+
+
+def matrices(*,
+             dims: _Sequence[str],
+             unit: _Union[_cpp.Unit, str] = _cpp.units.dimensionless,
+             values: _Union[_np.ndarray, list]):
+    """Constructs a :class:`Variable` with given dimensions holding an array
+    of 3x3 matrices.
+
+    :seealso: :py:func:`scipp.matrix`
+
+    :param dims: Dimension labels.
+    :param values: Initial values.
+    :param unit: Optional, data unit. Default=dimensionless
+    """
+    return _cpp.matrices(dims=dims, unit=unit, values=_to_eigen_layout(values))
+
+
+def vector(*,
+           unit: _Union[_cpp.Unit, str] = _cpp.units.dimensionless,
+           value: _Union[_np.ndarray, list]):
+    """Constructs a zero dimensional :class:`Variable` holding a single length-3
+    vector.
+
+    :seealso: :py:func:`scipp.vectors`
+
+    :param value: Initial value, a list or 1-D numpy array.
+    :param unit: Optional, unit. Default=dimensionless
+    :returns: A scalar (zero-dimensional) Variable.
+    :rtype: Variable
+    """
+    return _cpp.vectors(dims=[], unit=unit, values=value)
+
+
+def vectors(*,
+            dims: _Sequence[str],
+            unit: _Union[_cpp.Unit, str] = _cpp.units.dimensionless,
+            values: _Union[_np.ndarray, list]):
+    """Constructs a :class:`Variable` with given dimensions holding an array
+    of length-3 vectors.
+
+    :seealso: :py:func:`scipp.vector`
+
+    :param dims: Dimension labels.
+    :param values: Initial values.
+    :param unit: Optional, data unit. Default=dimensionless
+    """
+    return _cpp.vectors(dims=dims, unit=unit, values=values)
+
+
 def empty_like(var: _cpp.Variable) -> _cpp.Variable:
     """Constructs a :class:`Variable` with the same dims, shape, unit and dtype
     as the input variable, but with uninitialized values. If the input