+++ python/tests/plotting/test_plot_3d.py
@@ -21,12 +21,9 @@ def make_data_with_position_vectors():
     a = np.arange(M * N).reshape([M, N]) * np.sin(y)
     d = sc.Dataset()
     d['a'] = sc.Variable(['tof', 'xyz'], values=a)
-    d.coords['xyz'] = sc.Variable(['xyz'],
-                                  values=np.array([x, y, z]).T,
-                                  dtype=sc.dtype.vector_3_float64)
-    d.coords['pos'] = sc.Variable(['xyz'],
-                                  values=np.array([x, y, z]).T + 20.0,
-                                  dtype=sc.dtype.vector_3_float64)
+    d.coords['xyz'] = sc.vectors(dims=['xyz'], values=np.array([x, y, z]).T)
+    d.coords['pos'] = sc.vectors(dims=['xyz'],
+                                 values=np.array([x, y, z]).T + 20.0)
     d.coords['tof'] = sc.Variable(['tof'], values=tof)
     return d
 
@@ -110,15 +107,10 @@ def test_plot_3d_with_2d_position_coordinate():
         data=sc.Variable(['x', 'y', 'tof'],
                          values=np.arange(nx * ny *
                                           ntof).reshape(nx, ny, ntof)))
-    da.coords['pos'] = sc.fold(sc.Variable(
-        ['xyz'],
-        values=np.array([xx, yy, np.zeros_like(xx)]).T.reshape(nx * ny, 3),
-        dtype=sc.dtype.vector_3_float64),
-                               dim='xyz',
-                               sizes={
-                                   'x': nx,
-                                   'y': ny
-                               })
+    da.coords['pos'] = sc.vectors(dims=['x', 'y'],
+                                  values=np.array([xx, yy,
+                                                   np.zeros_like(xx)
+                                                   ]).T.reshape(nx, ny, 3))
     da.coords['tof'] = sc.Variable(['tof'],
                                    values=np.arange(ntof + 1,
                                                     dtype=np.float64))